[REFACTOR]: rename smtp to smtp_provider
