+++ internal/smtp_provider/net_test.go
@@ -1,4 +1,4 @@
-package smtp
+package smtp_provider
 
 import (
 	"errors"
